Slightly cleaner version of uploader
Summary:
I was having issues where I wasn't supplying all the parameters to the
uploader builder, so I decided to make it harder to get it wrong.

I also cleaned up the naming a bit.

I'm not a huge fan of this pattern at all in go, but I don't want to
make bigger changes right now

Test Plan: It's been integration tested inside spade, need to update spade_edge and others to use this

Reviewers: maasted, mperrone

Reviewed By: mperrone

Subscribers: dgarg, rhu

Differential Revision: https://phabricator.internal.justin.tv/D3171