aws_utils: add rollbar hook to logrus logger
Summary:
Add rollbar hook to logrus. Logrus assumes that the hooks attached to it are
synchronous, so when you call Fatal/Panic, it will print the log message, call
the hooks, and then immedately panic or fatal, exiting the program before the
async rollbar client can actually get the messages to rollbar. To get around
that, I modified the logrus wrapper to have fatals and panics be sent to logrus
as Errors, wait for the rollbar message to be sent, and then fatal/panic.
Since we duplicate the log level gate in our wrapper, it funtions identically
to basic logrus except for the log message printed out has "level":"error".
Also Rollbar won't get critical levels for those errors.

I considered not using the hook at all and making the rollbar calls in the
wrapper, but the logrus.Entry doesn't get populated with the message, level, and
time until the entry is logged.

A problem I didn't solve was that if you call logger.Panic somewhere, it will
get logged, sent to rollbar, and then a panic() will happen. The defer
logger.LogPanic() will catch it, and then log it again. I didn't know how to not
have that duplicated but still have all other panics not intentionally generated
get logged. Input requested.

Test Plan: ran with modified dbstorer to error, fatal, panic. saw logs outputted and messages in rollbar,

Reviewers: clgroft, dwe

Reviewed By: dwe

Subscribers: clgroft, #scieng

Differential Revision: https://phabricator.internal.justin.tv/D3844